move Word/Digest conversions to LeafIndex
